--- digiwf-engine/digiwf-engine-service/src/main/java/io/muenchendigital/digiwf/shared/configuration/SecurityCamundaConfig.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik der Landeshauptstadt München, 2020
- */
-
-package io.muenchendigital.digiwf.shared.configuration;
-
-import io.muenchendigital.digiwf.legacy.user.domain.service.UserService;
-import io.muenchendigital.digiwf.shared.security.CamundaUserAuthenticationFilter;
-import io.muenchendigital.digiwf.shared.security.UserAuthenticationProvider;
-import lombok.RequiredArgsConstructor;
-import org.camunda.bpm.engine.IdentityService;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-
-/**
- * Camunda Security configuration.
- * Adds the corresponding filter.
- *
- * @author externer.dl.horn
- */
-@Configuration
-@Profile("!no-security")
-@RequiredArgsConstructor
-public class SecurityCamundaConfig {
-
-    private final IdentityService identityService;
-    private final UserAuthenticationProvider userProvider;
-    private final UserService userService;
-
-    @Bean
-    public FilterRegistrationBean statelessUserAuthenticationFilter() {
-        final FilterRegistrationBean filterRegistration = new FilterRegistrationBean();
-        filterRegistration.setFilter(new CamundaUserAuthenticationFilter(this.userProvider, this.userService, this.identityService));
-        filterRegistration.setOrder(102); // make sure the filter is registered after the Spring Security Filter Chain
-        filterRegistration.addUrlPatterns("/rest/*");
-        return filterRegistration;
-    }
-
-}
